requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0
python-dotenv==1.0.0
ruff>=0.5.0

opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0
opentelemetry-instrumentation-httpx==0.41b0
opentelemetry-exporter-otlp==1.20.0
prometheus-fastapi-instrumentator==6.1.0